improvement on chunking

+++ drex/utils/reliability/utils.py
@@ -113,7 +113,7 @@ def build_building_blocks(m,n,p):
             row.append(elt)
             elt=(elt*a)%p
         building_blocks.append(row)
-    building_blocks_array = np.array(building_blocks, dtype=np.uint16)
+    building_blocks_array = np.array(building_blocks, dtype=np.uint8)
     return building_blocks_array
     
 def elementary_symmetric_functions(m,L,p):
